# openapi-power-automate.yaml
swagger: '2.0'
info:
  title: SLH TMS API
  version: '1.2'
host: slh-tms-api-prod.gentlepond-08dba66b.uksouth.azurecontainerapps.io
basePath: /api/v1
schemes: [https]
produces: [application/json]
consumes: [application/json]
securityDefinitions:
  oauth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://login.microsoftonline.com/REPLACE_TENANT_ID/oauth2/v2.0/authorize
    tokenUrl: https://login.microsoftonline.com/REPLACE_TENANT_ID/oauth2/v2.0/token
    scopes:
      api://REPLACE_API_CLIENT_ID/Tms.Access: Access the TMS API
security: [{ oauth2: [api://REPLACE_API_CLIENT_ID/Tms.Access] }]
paths:
...
  /staging:
    get:
      operationId: ListStagedImports
      summary: List staged records for review
      parameters:
        - { name: status, in: query, type: string, required: false, enum: [PendingReview, Approved, Rejected, Promoted, Failed] }
        - { name: entityType, in: query, type: string, required: false }
        - { name: take, in: query, type: integer, required: false, default: 100, maximum: 2000 }
      responses: { '200': { description: Staged records } }
    post:
      operationId: SubmitStagedImport
      summary: Submit an import for review
      parameters:
        - name: body
          in: body
          required: true
          schema: { $ref: '#/definitions/StageImportRequest' }
      responses:
        '202': { description: Accepted for review, schema: { $ref: '#/definitions/StageImportResponse' } }
        '200': { description: Existing idempotent submission, schema: { $ref: '#/definitions/StageImportResponse' } }
  /staging/{id}:
    get:
      operationId: GetStagedImport
      summary: Get review status
      parameters: [{ name: id, in: path, required: true, type: string, format: uuid }]
      responses: { '200': { description: Staged record } }
  /staging/{id}/approve:
    post:
      operationId: ApproveStagedImport
      summary: Approve and promote a staged import
      parameters:
        - { name: id, in: path, required: true, type: string, format: uuid }
        - { name: body, in: body, required: true, schema: { $ref: '#/definitions/ReviewRequest' } }
      responses: { '200': { description: Promoted record } }
  /staging/{id}/reject:
    post:
      operationId: RejectStagedImport
      summary: Reject a staged import
      parameters:
        - { name: id, in: path, required: true, type: string, format: uuid }
        - { name: body, in: body, required: true, schema: { $ref: '#/definitions/ReviewRequest' } }
      responses: { '200': { description: Rejected record } }
...
  /staging/{id}/history:
    get:
      operationId: GetStagedImportHistory
      summary: Get append-only SQL payload and review history for a staged import
      parameters: [{ name: id, in: path, required: true, type: string, format: uuid }]
      responses:
        '200':
          description: Ordered staging history
          schema: { type: array, items: { $ref: '#/definitions/StagedImportHistoryEvent' } }
  /staging/pending/archive:
    post:
      operationId: ArchivePendingStagedImports
      summary: Archive pending staging rows without deleting their SQL history
      parameters:
        - { name: body, in: body, required: true, schema: { $ref: '#/definitions/ArchivePendingRequest' } }
      responses: { '200': { description: Pending rows archived and retained } }
...
definitions:
...
  StageImportRequest:
    type: object
    required: [entityType, idempotencyKey, payload]
    properties:
      entityType: { type: string, enum: [customer, vehicle, driver, trailer, site, marketcontact, order] }
      idempotencyKey: { type: string, example: email-message-id-or-source-key }
      source: { type: string, example: PowerAutomate/InfoMailbox }
      payload: { type: object, additionalProperties: true }
  StageImportResponse:
    type: object
    properties:
      stagingId: { type: string, format: uuid }
      status: { type: string, example: PendingReview }
      receivedAtUtc: { type: string, format: date-time }
      reviewUrl: { type: string }
...
  ArchivePendingRequest:
    type: object
    required: [reason]
    properties:
      reason: { type: string }
  StagedImportHistoryEvent:
    type: object
    properties:
      id: { type: string, format: uuid }
      stagedImportId: { type: string, format: uuid }
      eventType: { type: string, enum: [Received, Amended, Approved, Rejected, Promoted, Failed, Superseded, Archived] }
      previousStatus: { type: string, enum: [PendingReview, Approved, Rejected, Promoted, Failed, Archived] }
      newStatus: { type: string, enum: [PendingReview, Approved, Rejected, Promoted, Failed, Archived] }
      payload: { type: object, additionalProperties: true }
      note: { type: string }
      actor: { type: string }
      occurredAtUtc: { type: string, format: date-time }
...
  ReviewRequest:
    type: object
    properties: { note: { type: string } }
...